rewrite s:index_message
It makes more use of g:indexed_search_shortmess (it was used only in one
place before).  It also returns a value regardless of v:errmsg. If this
is undesired, it can be easily fixed.  It also returns [hl, msg] instead
of setting script variables.  s:ShowCurrentSearchIndex does that now
thought that will dealt with too.