remove the cmd stuff from s:ShowCurrentSearchIndex
It didn't have much reason to exist anymore.  It was probably a remnant
of an older design.  The only path that was ever taken was the one where
cmd == '!'.  In that case it print vim's error message if there was one
but it changed the highlight group and it was only displayed briefly.  I
could add that path back if someone wants it.
I didn't remove the function argument yet.