fix n/N
We try to not change the "view" (the position in the window), with the
implementation copied from vim-slash.  Also, try to accomodate those
that want to move with * in terms of unfolding.  Previously, no
unfolding was done on *.

The better way to "dont_move" would be to use CTRL-O to preserve the
original column of the cursor as well.  However, that kind of brakes the
plugin, since, when the cursor is not on the start of a word it
displays "between matches 3-4 of 10" and so-on.