Change to a different algorithm (fixes #20)
This algorithm is correct since it uses the same mechanism that the user
does, that is, the 'n' and 'N' keys and it's base case is faster than
our previous algorithm.  However, the performance related options of the
plugin are not (yet) enforced so things may be slower (for now) on large
files or with many matches.